Add velocity ramps (WIP)
Added ability to press on two note pads to set a velocity ramp for the notes between the pads between the velocity values of the start and ending note

For this to work you need to select a first note and then select a second note and it will adjust the velocities values of the all notes in between.

Still some work to be done on this (display related and using floats instead of integers for more accurate ramp calculations), but it is functioning.