Update submenu rendering for toggle's and automatable param's (#2345)
* Update Menu UI for interacting with toggle menu's

Updated UI for interacting with toggle (ON/OFF) menu's on OLED and 7SEG
- OLED renders a checkbox that shows current ON/OFF status. Selecting that menu with select encoder will toggle the checkbox as opposed to entering the menu
- 7SEG renders a dot at the end of the menu item to show current ON/OFF status. Selecting that menu with select encoder will toggle the dot as opposed to entering the menu.

Submenu's are also now rendered with a ">" at the end to indicate that it is a submenu.

There are a couple edge case menu's (e.g. CV / Gate Output, Community Features) that do not inherit these changes but I think those can be addressed as part of a separate PR as they are built differently (don't use toggle or submenu classes).

* Render Automatable Param Values in Submenu's

Submenu's for editing Unpatched and Patched param values now display the current value at the end of each submenu line item

* Fix PR links