Fix where arranger playback restarts
When you enter a clip that is extended out in arranger, and restart playback from within that clip, it will always restart that clip from the clip head in arranger, regardless of if the arranger scroll position is scrolled passed the clip head.

This fixes it so that it will start that clip from the scroll position if the scroll position is past the clip head.