Add midi device definition file support
Added ability to save and load midi cc labels as part of the song and midi presets.

Added ability to set midi cc labels on the deluge via automation view.

While in automation view, select a midi cc, hold a gold encoder and press the "name" shortcut to set the name for the midi cc.

Cleaned up code for the other rename UI classes also