Refactor Midi Follow CC assignments to use lookup tables and add rest of automateable parameters (#3257)
* Refactor parameter names so none contain dots

* Refactor midi follow to use direct assignment lookup tables

* format

* Fixed crash

* Fixes

* Docs

* docs

* docs

* docs

* Updated deluge midi follow loopy pro template

* Fix bug

* Docs

* add new envelopes to midi follow

* Reorder by CC

Update midi follow feedback to check context (#3387)

When midi feedback is sent while editing params with gold knobs or with select encoder while in the menu, it should check the context to ensure that feedback is not sent for params that do not match the current midi follow controllable context.

Update midi follow feedback to check context

When midi feedback is sent while editing params with gold knobs or with select encoder while in the menu, it should check the context to ensure that feedback is not sent for params that do not match the current midi follow controllable context.