Update autoParam / paramNodes to use valueOffset
The AutoParam class has been updated so that it is not overriding the current param value when editing / processing automation.

Instead, when editing automation (e.g. by recording or using automation view) or when automation is played back, a "currentValueOffset" variable is used to adjust the current param value.

Thus when you delete automation, all you need to do is reset the currentValueOffset to 0 and you end up back at the current param value.

Update param nodes to use value offset

Updated param nodes to store a value offset from the current auto param value instead of the full param value at the node.

The node value offset will then added to the current value to get the value at a node's position.