Pulse Selected Clip in Grid View (#2825)
* Pulse Selected Clip in Grid View

The selected clip in `Song Grid View` (which is the last clip you interacted with), now shows a `pulsing / breathing` animation whereby the clip clip will faded in and out between it's clip colour and a blurred version of the clip's colour.

* Update CHANGELOG.md

* Addressing minor comments

* Refactor pulsing colour calculations

* Fix color stepping issue

---------

add finally.h

Increased grid view pad brightness (#3896)

Removed grid view pad brightness adjustment which was added when the "selected clip pulsing" feature was added (to add greater focus on the selected clip).

I don't think it's absolutely necessary to do that, and since we've received complaints about the brightness, I am removing the adjustment

Co-Authored-By: Katherine Whitlock <kate@skylinesynths.nyc>